LayoutTests/http/tests/cookies/document-cookie-after-showModalDialog-expected.txt

CONSOLE MESSAGE: Window 'showModalDialog' function is deprecated and will be removed soon.
CONSOLE MESSAGE: showModalDialog() is deprecated and will be removed. Please use the <dialog> element instead.
Tests that document.cookie returns the right value after a showModalDialog() call

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS normalizeCookie(document.cookie) is "testKey=testValue"
PASS normalizeCookie(document.cookie) is "foo=bar; testKey=testValue"
PASS successfullyParsed is true

TEST COMPLETE
